fix :remove warning when building

+++ riscv/insns/mulqr.h
@@ -1,6 +1,6 @@
 require_extension('P');
 require_rv32;
-if ((RS1 != INT32_MIN) | (RS2 != INT32_MIN)) {
+if (((int32_t)RS1 != INT32_MIN) | ((int32_t)RS2 != INT32_MIN)) {
     WRITE_RD((((RS1 * RS2) >> 30) + 1) >> 1);
   } else {
     WRITE_RD(INT32_MAX);
